+++ stroom-app/src/main/java/stroom/startup/App.java
@@ -50,6 +50,7 @@ import stroom.dictionary.server.DictionaryResource;
 import stroom.dictionary.server.DictionaryStore;
 import stroom.dictionary.shared.DictionaryDoc;
 import stroom.dictionary.spring.DictionaryConfiguration;
+import stroom.dispatch.server.XsrfTokenServiceServlet;
 import stroom.dispatch.shared.DispatchService;
 import stroom.entity.server.SpringRequestFactoryServlet;
 import stroom.entity.server.util.ConnectionUtil;
@@ -149,6 +150,7 @@ public class App extends Application<Config> {
     private static final String SESSION_LIST_PATH = ResourcePaths.ROOT_PATH + "/sessionList";
     private static final String STATUS_PATH = ResourcePaths.ROOT_PATH + "/status";
     private static final String UI_PATH = ResourcePaths.ROOT_PATH + "/ui";
+    private static final String XSRF_TOKEN_RPC_PATH = ResourcePaths.ROOT_PATH + "/xsrf";
 
     private static String configPath;
 
@@ -376,6 +378,7 @@ public class App extends Application<Config> {
         SpringUtil.addServlet(servletContextHandler, applicationContext, RemoteFeedServiceRPC.class, REMOTING_RPC_PATH);
         SpringUtil.addServlet(servletContextHandler, applicationContext, DataFeedServlet.class, DATAFEED_PATH);
         SpringUtil.addServlet(servletContextHandler, applicationContext, DataFeedServlet.class, DATAFEED_PATH + "/*");
+        SpringUtil.addServlet(servletContextHandler, applicationContext, XsrfTokenServiceServlet.class, XSRF_TOKEN_RPC_PATH);
 
         // Add session listeners.
         SpringUtil.addServletListener(environment.servlets(), applicationContext, SessionListListener.class);
@@ -411,6 +414,7 @@ public class App extends Application<Config> {
                 .put(makeBypassAuthInitParam(ECHO_PATH, false))
                 .put(makeBypassAuthInitParam(DEBUG_PATH, false))
                 .put(makeBypassAuthInitParam(REMOTING_RPC_PATH, false))
+                .put(makeBypassAuthInitParam(XSRF_TOKEN_RPC_PATH, false))
                 .build();
 
         SpringUtil.addFilter(servletContextHandler, applicationContext, SecurityFilter.class, "/*", initParams);